# perf: improve plan for proto_get_trial_plus.sql