--- a/master/static/migrations/20220504154053_add-checkpoints-v2-steps-completed-index.tx.down.sql
+++ b/master/static/migrations/20220504154053_add-checkpoints-v2-steps-completed-index.tx.down.sql
@@ -1,0 +1,2 @@
+-- This change wasn't caught on a version boundary; if a version downgrade is run, the down
+-- migration prior to us will also run, and it runs a delete cascade that will drop all our stuff.