## Description
Ran into some fairly serious perf issues without this index due to joins, wheres and sorts using this fields.
<!---
Lead with the intended commit body in this description field. For breaking changes,
please include "BREAKING CHANGE:" at the beginning of your commit body.
At a minimum, this section should include a bracketed reference to the Jira ticket,
e.g. "[DET-1234]". When squash-and-merging, copy this directly into the description field.
-->


## Test Plan
- [x] test for perf improvement
<!---
Describe the situations in which you've tested your change, and/or a screenshot as appropriate.
Reviewers may ask questions about this test plan to ensure adequate manual coverage of changes.
-->


## Commentary (optional)
Before
```
   QUERY PLAN                                                                                                                                                                                                                                                         
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Nested Loop Left Join  (cost=281552.23..283718.73 rows=1 width=532) (actual time=4163.052..4184.213 rows=1 loops=1)
   Join Filter: (bc.trial_id = searcher_info.trial_id)
   CTE searcher_info
     ->  Nested Loop  (cost=2.33..782.42 rows=100 width=40) (actual time=0.108..0.114 rows=1 loops=1)
           ->  Nested Loop  (cost=2.06..741.33 rows=100 width=8) (actual time=0.042..0.046 rows=1 loops=1)
                 ->  HashAggregate  (cost=1.77..2.77 rows=100 width=4) (actual time=0.013..0.015 rows=1 loops=1)
                       Group Key: unnest('{24348}'::integer[])
                       ->  ProjectSet  (cost=0.00..0.52 rows=100 width=4) (actual time=0.006..0.009 rows=1 loops=1)
                             ->  Result  (cost=0.00..0.01 rows=1 width=0) (actual time=0.001..0.002 rows=1 loops=1)
                 ->  Index Scan using trials_pkey on trials t_3  (cost=0.29..7.38 rows=1 width=8) (actual time=0.024..0.024 rows=1 loops=1)
                       Index Cond: (id = (unnest('{24348}'::integer[])))
           ->  Index Scan using experiments_pkey on experiments e_2  (cost=0.28..0.39 rows=1 width=822) (actual time=0.008..0.008 rows=1 loops=1)
                 Index Cond: (id = t_3.experiment_id)
           SubPlan 1
             ->  Result  (cost=0.00..0.02 rows=1 width=4) (actual time=0.016..0.016 rows=1 loops=1)
   CTE trial_validations
     ->  Nested Loop  (cost=0.29..954.11 rows=309 width=167) (actual time=0.025..0.027 rows=1 loops=1)
           ->  CTE Scan on searcher_info searcher_info_1  (cost=0.00..2.00 rows=100 width=40) (actual time=0.001..0.001 rows=1 loops=1)
           ->  Index Scan using ix_validations_trial_id on raw_validations raw_validations_2  (cost=0.29..9.44 rows=3 width=159) (actual time=0.013..0.013 rows=1 loops=1)
                 Index Cond: (trial_id = searcher_info_1.trial_id)
                 Filter: ((NOT archived) AND (state = 'COMPLETED'::validation_state) AND (((metrics -> 'validation_metrics'::text) ->> searcher_info_1.metric_name) IS NOT NULL))
   CTE best_validation
     ->  Hash Join  (cost=29.03..31.42 rows=1 width=112) (actual time=0.090..0.094 rows=1 loops=1)
           Hash Cond: (searcher_info_2.trial_id = v.trial_id)
           ->  CTE Scan on searcher_info searcher_info_2  (cost=0.00..2.00 rows=100 width=4) (actual time=0.000..0.001 rows=1 loops=1)
           ->  Hash  (cost=29.00..29.00 rows=2 width=52) (actual time=0.054..0.057 rows=1 loops=1)
                 Buckets: 1024  Batches: 1  Memory Usage: 9kB
                 ->  Subquery Scan on v  (cost=18.96..29.00 rows=2 width=52) (actual time=0.051..0.054 rows=1 loops=1)
                       Filter: (v.rank = 1)
                       ->  WindowAgg  (cost=18.96..25.14 rows=309 width=68) (actual time=0.049..0.051 rows=1 loops=1)
                             ->  Sort  (cost=18.96..19.73 rows=309 width=60) (actual time=0.040..0.041 rows=1 loops=1)
                                   Sort Key: v_1.trial_id, v_1.signed_searcher_metric
                                   Sort Method: quicksort  Memory: 25kB
                                   ->  CTE Scan on trial_validations v_1  (cost=0.00..6.18 rows=309 width=60) (actual time=0.028..0.029 rows=1 loops=1)
   CTE latest_validation
     ->  Hash Join  (cost=29.03..31.42 rows=1 width=112) (actual time=0.038..0.044 rows=1 loops=1)
           Hash Cond: (searcher_info_3.trial_id = v_2.trial_id)
           ->  CTE Scan on searcher_info searcher_info_3  (cost=0.00..2.00 rows=100 width=4) (actual time=0.001..0.002 rows=1 loops=1)
           ->  Hash  (cost=29.00..29.00 rows=2 width=52) (actual time=0.021..0.024 rows=1 loops=1)
                 Buckets: 1024  Batches: 1  Memory Usage: 9kB
                 ->  Subquery Scan on v_2  (cost=18.96..29.00 rows=2 width=52) (actual time=0.018..0.021 rows=1 loops=1)
                       Filter: (v_2.rank = 1)
                       ->  WindowAgg  (cost=18.96..25.14 rows=309 width=68) (actual time=0.016..0.018 rows=1 loops=1)
                             ->  Sort  (cost=18.96..19.73 rows=309 width=52) (actual time=0.013..0.014 rows=1 loops=1)
                                   Sort Key: v_3.trial_id, v_3.end_time DESC
                                   Sort Method: quicksort  Memory: 25kB
                                   ->  CTE Scan on trial_validations v_3  (cost=0.00..6.18 rows=309 width=52) (actual time=0.001..0.002 rows=1 loops=1)
   CTE best_checkpoint
     ->  Subquery Scan on c_2  (cost=125755.38..125760.85 rows=1 width=112) (actual time=1844.831..1856.115 rows=1 loops=1)
           Filter: (c_2.rank = 1)
           ->  WindowAgg  (cost=125755.38..125758.74 rows=168 width=336) (actual time=1844.823..1856.105 rows=1 loops=1)
                 ->  Sort  (cost=125755.38..125755.80 rows=168 width=80) (actual time=1844.813..1856.093 rows=1 loops=1)
                       Sort Key: v_4.trial_id, v_4.signed_searcher_metric
                       Sort Method: quicksort  Memory: 25kB
                       ->  Hash Join  (cost=122763.19..125749.17 rows=168 width=80) (actual time=1844.644..1856.077 rows=1 loops=1)
                             Hash Cond: (("*SELECT* 1_1".steps_completed = v_4.total_batches) AND ("*SELECT* 1_1".trial_id = v_4.trial_id))
                             ->  Unique  (cost=122752.37..123729.05 rows=21704 width=316) (actual time=1765.224..1851.322 rows=21449 loops=1)
                                   ->  Sort  (cost=122752.37..122806.63 rows=21704 width=316) (actual time=1765.221..1828.297 rows=21449 loops=1)
                                         Sort Key: (("*SELECT* 1_1".id)::bigint), "*SELECT* 1_1".uuid, "*SELECT* 1_1".task_id, "*SELECT* 1_1".allocation_id, "*SELECT* 1_1".report_time, "*SELECT* 1_1".state, "*SELECT* 1_1".resources, "*SELECT* 1_1".metadata, "*SELECT* 1_1".trial_id, "*SELECT* 1_1".experiment_id, "*SELECT* 1_1".experiment_config, "*SELECT* 1_1".hparams, "*SELECT* 1_1".training_metrics, "*SELECT* 1_1".validation_metrics, "*SELECT* 1_1".searcher_metric, "*SELECT* 1_1".steps_completed, (1)
                                         Sort Method: external merge  Disk: 118400kB
                                         ->  Append  (cost=21016.37..117997.07 rows=21704 width=316) (actual time=317.043..1279.272 rows=21449 loops=1)
                                               ->  Subquery Scan on "*SELECT* 1_1"  (cost=21016.37..112649.67 rows=21614 width=2862) (actual time=317.041..1243.336 rows=21357 loops=1)
                                                     ->  Gather  (cost=21016.37..112379.50 rows=21614 width=2858) (actual time=317.038..1233.187 rows=21357 loops=1)
                                                           Workers Planned: 1
                                                           Workers Launched: 1
                                                           ->  Hash Left Join  (cost=20016.37..109218.10 rows=12714 width=2858) (actual time=316.245..1234.277 rows=10678 loops=2)
                                                                 Hash Cond: (t_4.experiment_id = e_3.id)
                                                                 ->  Nested Loop Left Join  (cost=19437.01..108192.07 rows=12714 width=2115) (actual time=313.647..460.773 rows=10678 loops=2)
                                                                       Filter: ((s_2.archived IS NULL) OR (NOT s_2.archived))
                                                                       ->  Parallel Hash Left Join  (cost=19436.58..22452.09 rows=12801 width=850) (actual time=313.594..340.221 rows=10674 loops=2)
                                                                             Hash Cond: (c_3.trial_id = t_4.id)
                                                                             ->  Merge Left Join  (cost=15910.13..16572.03 rows=12801 width=509) (actual time=163.707..235.877 rows=10674 loops=2)
                                                                                   Merge Cond: ((c_3.trial_id = raw_validations_3.trial_id) AND (c_3.total_batches = raw_validations_3.total_batches))
                                                                                   ->  Sort  (cost=4911.41..4943.42 rows=12801 width=370) (actual time=37.272..43.798 rows=10674 loops=2)
                                                                                         Sort Key: c_3.trial_id, c_3.total_batches
                                                                                         Sort Method: external merge  Disk: 4848kB
                                                                                         Worker 0:  Sort Method: external merge  Disk: 3152kB
                                                                                         ->  Parallel Seq Scan on raw_checkpoints c_3  (cost=0.00..1847.13 rows=12801 width=370) (actual time=0.013..9.701 rows=10674 loops=2)
                                                                                               Filter: (state = 'COMPLETED'::checkpoint_state)
                                                                                               Rows Removed by Filter: 3682
                                                                                   ->  Materialize  (cost=10998.72..11273.18 rows=54893 width=147) (actual time=126.410..163.213 rows=55776 loops=2)
                                                                                         ->  Sort  (cost=10998.72..11135.95 rows=54893 width=147) (actual time=126.402..150.738 rows=55776 loops=2)
                                                                                               Sort Key: raw_validations_3.trial_id, raw_validations_3.total_batches
                                                                                               Sort Method: external merge  Disk: 8488kB
                                                                                               Worker 0:  Sort Method: external merge  Disk: 8488kB
                                                                                               ->  Seq Scan on raw_validations raw_validations_3  (cost=0.00..2547.45 rows=54893 width=147) (actual time=0.040..28.815 rows=55916 loops=2)
                                                                                                     Filter: (NOT archived)
                                                                                                     Rows Removed by Filter: 247
                                                                             ->  Parallel Hash  (cost=2703.31..2703.31 rows=14331 width=341) (actual time=21.121..21.121 rows=12192 loops=2)
                                                                                   Buckets: 16384  Batches: 4  Memory Usage: 2464kB
                                                                                   ->  Parallel Seq Scan on trials t_4  (cost=0.00..2703.31 rows=14331 width=341) (actual time=0.014..7.479 rows=12192 loops=2)
                                                                       ->  Index Scan using steps_trial_id_run_id_total_batches_unique on raw_steps s_2  (cost=0.42..6.69 rows=1 width=1278) (actual time=0.009..0.010 rows=1 loops=21349)
                                                                             Index Cond: ((c_3.trial_id = trial_id) AND (c_3.total_batches = total_batches))
                                                                 ->  Hash  (cost=559.16..559.16 rows=1616 width=822) (actual time=2.450..2.450 rows=1618 loops=2)
                                                                       Buckets: 2048  Batches: 1  Memory Usage: 1368kB
                                                                       ->  Seq Scan on experiments e_3  (cost=0.00..559.16 rows=1616 width=822) (actual time=0.019..1.287 rows=1618 loops=2)
                                               ->  Nested Loop Left Join  (cost=25.94..5237.98 rows=90 width=3285) (actual time=0.349..31.548 rows=92 loops=1)
                                                     Join Filter: (((c_4.metadata ->> 'steps_completed'::text))::integer = raw_validations_4.total_batches)
                                                     Rows Removed by Join Filter: 62
                                                     ->  Nested Loop Left Join  (cost=25.65..4888.66 rows=90 width=3237) (actual time=0.270..28.937 rows=92 loops=1)
                                                           ->  Nested Loop Left Join  (cost=25.37..4640.51 rows=90 width=2419) (actual time=0.235..28.587 rows=92 loops=1)
                                                                 Filter: ((s_3.archived IS NULL) OR (NOT s_3.archived))
                                                                 ->  Hash Right Join  (cost=24.94..3804.00 rows=91 width=1150) (actual time=0.203..21.577 rows=92 loops=1)
                                                                       Hash Cond: (t_5.task_id = c_4.task_id)
                                                                       ->  Seq Scan on trials t_5  (cost=0.00..2803.63 rows=24363 width=341) (actual time=0.011..14.781 rows=24383 loops=1)
                                                                       ->  Hash  (cost=23.80..23.80 rows=91 width=827) (actual time=0.167..0.168 rows=92 loops=1)
                                                                             Buckets: 1024  Batches: 1  Memory Usage: 98kB
                                                                             ->  Seq Scan on checkpoints_v2 c_4  (cost=0.00..23.80 rows=91 width=827) (actual time=0.015..0.117 rows=92 loops=1)
                                                                                   Filter: (state = 'COMPLETED'::checkpoint_state)
                                                                                   Rows Removed by Filter: 52
                                                                 ->  Index Scan using steps_trial_id_run_id_total_batches_unique on raw_steps s_3  (cost=0.43..9.18 rows=1 width=1278) (actual time=0.074..0.074 rows=1 loops=92)
                                                                       Index Cond: ((t_5.id = trial_id) AND (((c_4.metadata ->> 'steps_completed'::text))::integer = total_batches))
                                                           ->  Index Scan using experiments_pkey on experiments e_4  (cost=0.28..2.76 rows=1 width=822) (actual time=0.003..0.003 rows=1 loops=92)
                                                                 Index Cond: (t_5.experiment_id = id)
                                                     ->  Index Scan using ix_validations_trial_id on raw_validations raw_validations_4  (cost=0.29..3.80 rows=3 width=147) (actual time=0.003..0.004 rows=1 loops=92)
                                                           Index Cond: (t_5.id = trial_id)
                                                           Filter: (NOT archived)
                             ->  Hash  (cost=6.18..6.18 rows=309 width=16) (actual time=0.003..0.004 rows=1 loops=1)
                                   Buckets: 1024  Batches: 1  Memory Usage: 9kB
                                   ->  CTE Scan on trial_validations v_4  (cost=0.00..6.18 rows=309 width=16) (actual time=0.001..0.001 rows=1 loops=1)
   ->  Nested Loop Left Join  (cost=153992.00..155792.04 rows=1 width=456) (actual time=2317.305..2327.157 rows=1 loops=1)
         Join Filter: (lv.trial_id = searcher_info.trial_id)
         ->  Nested Loop Left Join  (cost=153992.00..155792.01 rows=1 width=428) (actual time=2317.257..2327.102 rows=1 loops=1)
               Join Filter: (bv.trial_id = searcher_info.trial_id)
               ->  Nested Loop  (cost=153992.00..155791.98 rows=1 width=400) (actual time=2317.155..2326.996 rows=1 loops=1)
                     Join Filter: (t.id = unnest.unnest)
                     ->  Function Scan on unnest  (cost=0.00..1.00 rows=100 width=12) (actual time=0.007..0.008 rows=1 loops=1)
                     ->  Materialize  (cost=153992.00..155641.22 rows=100 width=392) (actual time=2317.145..2326.983 rows=1 loops=1)
                           ->  Merge Right Join  (cost=153992.00..155640.72 rows=100 width=392) (actual time=2317.143..2326.979 rows=1 loops=1)
                                 Merge Cond: ((("*SELECT* 1".id)::bigint) = t.warm_start_checkpoint_id)
                                 ->  Unique  (cost=153248.18..154537.97 rows=28662 width=316) (actual time=2317.006..2326.837 rows=1 loops=1)
                                       ->  Sort  (cost=153248.18..153319.83 rows=28662 width=316) (actual time=2317.003..2326.833 rows=1 loops=1)
                                             Sort Key: (("*SELECT* 1".id)::bigint), "*SELECT* 1".uuid, "*SELECT* 1".task_id, "*SELECT* 1".allocation_id, "*SELECT* 1".report_time, "*SELECT* 1".state, "*SELECT* 1".resources, "*SELECT* 1".metadata, "*SELECT* 1".trial_id, "*SELECT* 1".experiment_id, "*SELECT* 1".experiment_config, "*SELECT* 1".hparams, "*SELECT* 1".training_metrics, "*SELECT* 1".validation_metrics, "*SELECT* 1".searcher_metric, "*SELECT* 1".steps_completed, (1)
                                             Sort Method: external merge  Disk: 159352kB
                                             ->  Append  (cost=21983.44..146912.21 rows=28662 width=316) (actual time=370.698..1618.561 rows=28866 loops=1)
                                                   ->  Subquery Scan on "*SELECT* 1"  (cost=21983.44..140697.62 rows=28519 width=2862) (actual time=370.696..1585.670 rows=28722 loops=1)
                                                         ->  Gather  (cost=21983.44..140341.13 rows=28519 width=2858) (actual time=370.693..1571.896 rows=28722 loops=1)
                                                               Workers Planned: 1
                                                               Workers Launched: 1
                                                               ->  Hash Left Join  (cost=20983.44..136489.23 rows=16776 width=2858) (actual time=370.183..1577.453 rows=14361 loops=2)
                                                                     Hash Cond: (t_1.experiment_id = e.id)
…